requirements.txt
Pillow >= 5.1.0